move testing files to seperate folders
