+++ Saucier720-app/src/app/testing/list/list.component.spec.cy.ts
@@ -1,5 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { ListComponent } from '../list/list.component';
+import { ListComponent } from '../../list/list.component';
 
 describe('ListComponent', () => {
   it('mounts', () => {
