+++ Saucier720-app/src/app/testing/recipes/recipes.component.spec.cy.ts
@@ -1,5 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { RecipesComponent } from '../recipes/recipes.component';
+import { RecipesComponent } from '../../recipes/recipes.component';
 
 describe('RecipesComponent', () => {
   it('mounts', () => {
